Complete signal handler
Note that this is currently not working as expected. For some reason, DynamoRIO doesn't perform a context switch whilst in the SIGSEGV handler and thus it doesn't allow the other thread to process.
In the next commit, I'll try to insert the signal handler as application code, since I tried running a binary that has its own signal handler with DynamoRIO (and without any client) and that seems to work fine.