Rewrite signal handler as DynamoRIO event
This is needed because inserting the signal handler as application code comes with some challenges.
If the signal handler was inserted as application code, when an instruction operating on the queue
fails (let its PC be y), the application will be presented with the previous PC of an origial
application instruction (let its PC be x). Thus, all instructions between x and y will be
rexecuted after the queue pointers are adjusted. The machine context must thus be reverted to where
it was when x was executed. If x is in another basic block, this further complicates the instrumentation.